Added functionality for multiple copy using single function
