+++ src/pages/TeamPage.jsx
@@ -70,7 +70,9 @@ export const TeamPage = () => {
 
         <div className='team-page-cta'>
           <div className='slack-cta'>
-            <h4>Join us, become a contributor!</h4>
+            <h4 style={{ textAlign: 'center' }}>
+              Join us, become a contributor!
+            </h4>
 
             <p>
               Be part of the team, join our Slack channel, and chat with our
